% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialstock_update.R
\name{materialstock_tm_bak}
\alias{materialstock_tm_bak}
\title{erp备份表}
\usage{
materialstock_tm_bak(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{

}
\description{
erp备份表
}
\examples{
materialstock_tm_bak()
}
