% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialstock_update.R
\name{materialstock_tm_delete}
\alias{materialstock_tm_delete}
\title{erp删除清空表}
\usage{
materialstock_tm_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{

}
\description{
erp删除清空表
}
\examples{
materialstock_tm_delete()
}
